social/requirements.txt
pyjwt==2.1.0
Django==3.0.5
djangorestframework==3.11.0
requests==2.23.0

